// packages/ioc/src/metadata/meta.d.ts
export interface PropertyMetadata<T = any> extends Omit<Parameter<T>, 'name'> {
    /**
     * property type
     *
     * @type {SymbolType}
     */
    type?: AbstractType<T>;
    /**
     * property name
     *
     * @type {string}
     */
    propertyKey: string;
}